+++ packages/expo-modules-core/android/src/main/java/expo/modules/interfaces/font/FontManagerInterface.java
@@ -1,7 +1,7 @@
-package org.unimodules.interfaces.font;
+package expo.modules.interfaces.font;
 
 import android.graphics.Typeface;
 
-public interface FontManager {
+public interface FontManagerInterface {
   void setTypeface(String fontFamilyName, int style, Typeface typeface);
 }